requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
Django==2.2.2
enum34==1.1.6
firebase==3.0.1
firebase-admin==2.17.0
futures==3.1.1
gcloud==0.17.0
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
httplib2==0.13.0
jws==0.1.3
msgpack==0.6.1
oauth2client==3.0.0
protobuf==3.9.0rc1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2019.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
